Clipboard works, need to refresh RecyclerView
Successfully got the most recent plaintext item on the clipboard added
to a list. Problem is, I can't get the recyclerview to refresh when it
is added. For now I have two buttons that add the copied item to the
list, and a non-functioning button that is supposed to refresh the
recyclerview.