Updated ArrayList<String> to ArrayList<Clip>, implemented date and star
1. Clip object is now used in the adapter. Contents, date, and favorite
status are accessed through get methods within the Clip object.

2. readFromClipboard now grabs the current date and time and puts it
through the adapter.

3. New method which readFromClipboard utilizes, checkDuplicate. Iterates
through all objects in clipList (which is what I renamed recentList to)
and makes sure the contents do not equal.

4. Favorite button listeners are now included in the adapter. If the
star is pressed, it will set the favorite boolean in the Clip object to
true or false respectively, and log it to console.

KNOWN BUGS: When you add another item to index 0 in the list by
refreshing, the favorite (and POSSIBLY date, unlikely but haven't
tested) will stay at their index instead of move down with the rest of
the stuff. This the favorite status to always be at those indexes,
instead of their corresponding Clip objects.